backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
requests==2.31.0
mrz==0.6.2
Pillow==10.2.0
numpy==1.26.3
opencv-python-headless==4.9.0.80
python-dotenv==1.0.0
openpyxl==3.1.2
python-jose[cryptography]==3.3.0
